+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- delete_raw to rest utils
+- verify method to Response class
 
 ## [0.2.1] - 2024-03-06
 ## Fixed